requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
